requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
pyserial==3.5
requests==2.32.3
urllib3==2.2.2